Started porting components from the MADE.UI.Controls platform specific libraries.
