+++ MADE App Components.sln
@@ -21,8 +21,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MADE.UI.Controls.iOS", "MAD
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MADE.UI.Controls.UWP", "MADE.UI.Controls.UWP\MADE.UI.Controls.UWP.csproj", "{2C70160A-FFDD-4532-9572-12398B1C0722}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MADE.UI.Android", "MADE.UI.Android\MADE.UI.Android.csproj", "{1319A64A-5DA2-475B-ACF8-2884F892A371}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Testing", "Testing", "{28A9006E-1F26-4A33-8290-7A43D8EE2A9D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Playground", "Playground", "{835CE285-784A-4121-AC03-1E0ABC6A8D10}"
@@ -355,54 +353,6 @@ Global
 		{2C70160A-FFDD-4532-9572-12398B1C0722}.Release|x64.Build.0 = Release|x64
 		{2C70160A-FFDD-4532-9572-12398B1C0722}.Release|x86.ActiveCfg = Release|x86
 		{2C70160A-FFDD-4532-9572-12398B1C0722}.Release|x86.Build.0 = Release|x86
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|ARM.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|ARM.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|iPhone.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|iPhone.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|iPhoneSimulator.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|x64.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|x64.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|x86.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Ad-Hoc|x86.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|Any CPU.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|ARM.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|ARM.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|iPhone.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|iPhone.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|iPhoneSimulator.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|x64.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|x64.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|x86.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.AppStore|x86.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|ARM.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|x64.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Debug|x86.Build.0 = Debug|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|ARM.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|ARM.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|iPhone.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|x64.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|x64.Build.0 = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|x86.ActiveCfg = Release|Any CPU
-		{1319A64A-5DA2-475B-ACF8-2884F892A371}.Release|x86.Build.0 = Release|Any CPU
 		{26BEFC2B-AF82-41FB-AF3E-E0C8F8243509}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
 		{26BEFC2B-AF82-41FB-AF3E-E0C8F8243509}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
 		{26BEFC2B-AF82-41FB-AF3E-E0C8F8243509}.Ad-Hoc|Any CPU.Deploy.0 = Release|Any CPU
@@ -1543,7 +1493,6 @@ Global
 		{2FC17F7C-A6A8-44AF-B578-A03B7BFA5B65} = {2DB08E33-E63B-480B-8DB9-9B21C1CE9D6F}
 		{C7E54ACD-8B36-4318-A60C-D7EDDA63A26D} = {2DB08E33-E63B-480B-8DB9-9B21C1CE9D6F}
 		{2C70160A-FFDD-4532-9572-12398B1C0722} = {2DB08E33-E63B-480B-8DB9-9B21C1CE9D6F}
-		{1319A64A-5DA2-475B-ACF8-2884F892A371} = {E1037749-607E-4D93-A35A-26A4BED58A59}
 		{835CE285-784A-4121-AC03-1E0ABC6A8D10} = {28A9006E-1F26-4A33-8290-7A43D8EE2A9D}
 		{DB8B159E-0B9B-4F08-9475-E7A4037AFE6B} = {28A9006E-1F26-4A33-8290-7A43D8EE2A9D}
 		{26BEFC2B-AF82-41FB-AF3E-E0C8F8243509} = {835CE285-784A-4121-AC03-1E0ABC6A8D10}
