+++ MADE.UI.Controls.Android/MADE.UI.Controls.Android.csproj
@@ -47,17 +47,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Control.cs" />
-    <Compile Include="Extensions\Extensions.Controls.cs" />
-    <Compile Include="Extensions\Extensions.LinearLayout.cs" />
     <Compile Include="IAndroidControl.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MADE.UI.Android\MADE.UI.Android.csproj">
-      <Project>{1319a64a-5da2-475b-acf8-2884f892a371}</Project>
-      <Name>MADE.UI.Android</Name>
-    </ProjectReference>
     <ProjectReference Include="..\MADE.UI.Controls\MADE.UI.Controls.csproj">
       <Project>{2469a729-c7f3-4a68-9d1b-f089951aadab}</Project>
       <Name>MADE.UI.Controls</Name>
