+++ MADE.UI.Controls.UWP/MADE.UI.Controls.UWP.csproj
@@ -109,7 +109,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Control.cs" />
-    <Compile Include="Extensions\Extensions.Controls.cs" />
     <Compile Include="IWindowsControl.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\MADE.Controls.UWP.rd.xml" />
@@ -128,6 +127,9 @@
       <Project>{49f44ac9-02d1-44b4-b98f-9949bb7d00a4}</Project>
       <Name>MADE.UI</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Extensions\" />
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
