+++ MADE.UI.Controls.iOS/MADE.UI.Controls.iOS.csproj
@@ -39,14 +39,13 @@
     <Reference Include="Xamarin.iOS" />
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="Extensions\" />
     <Folder Include="Resources\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Control.cs" />
-    <Compile Include="Extensions\Extensions.Controls.cs" />
     <Compile Include="IIOSControl.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Extensions\Extensions.UIStackView.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MADE.UI.Controls\MADE.UI.Controls.csproj">
