+++ MADE.UI.Controls/Extensions/Extensions.Android.Controls.cs
@@ -1,9 +1,10 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+﻿#if __ANDROID__
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="Extensions.Controls.cs" company="MADE Apps">
 //   Copyright (c) MADE Apps.
 // </copyright>
 // <summary>
-//   Defines a collection of extensions for UWP controls.
+//   Defines a collection of extensions for Android controls.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
@@ -11,11 +12,10 @@ namespace MADE.UI.Controls
 {
     using System;
 
-    using Windows.UI.Xaml;
-    using Windows.UI.Xaml.Controls;
+    using Android.Views;
 
     /// <summary>
-    /// Defines a collection of extensions for UWP controls.
+    /// Defines a collection of extensions for Android controls.
     /// </summary>
     public static partial class Extensions
     {
@@ -28,7 +28,7 @@ namespace MADE.UI.Controls
         /// <param name="isVisible">
         /// A value indicating whether the control is visible.
         /// </param>
-        public static void SetVisible(this UIElement view, bool isVisible)
+        public static void SetVisible(this View view, bool isVisible)
         {
             SetVisible(view, isVisible, false);
         }
@@ -46,24 +46,25 @@ namespace MADE.UI.Controls
         /// <param name="updateChildViews">
         /// A value indicating whether to update the control's child visibility states.
         /// </param>
-        public static void SetVisible(this UIElement view, bool isVisible, bool updateChildViews)
+        public static void SetVisible(this View view, bool isVisible, bool updateChildViews)
         {
             if (view == null)
             {
                 return;
             }
 
-            view.Visibility = isVisible ? Visibility.Visible : Visibility.Collapsed;
+            view.Visibility = isVisible ? ViewStates.Visible : ViewStates.Gone;
 
-            if (!updateChildViews || !(view is Panel panel))
+            if (!updateChildViews || !(view is ViewGroup viewGroup))
             {
                 return;
             }
 
-            foreach (UIElement child in panel.Children)
+            for (int i = 0; i < viewGroup.ChildCount; i++)
             {
                 try
                 {
+                    View child = viewGroup.GetChildAt(i);
                     child?.SetVisible(isVisible, true);
                 }
                 catch (Exception)
@@ -74,3 +75,4 @@ namespace MADE.UI.Controls
         }
     }
 }
+#endif
