+++ MADE.UI.Controls/Extensions/Extensions.Android.LinearLayout.cs
@@ -1,4 +1,5 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+﻿#if __ANDROID__
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="Extensions.LinearLayout.cs" company="MADE Apps">
 //   Copyright (c) MADE Apps.
 // </copyright>
@@ -42,3 +43,4 @@ namespace MADE.UI.Controls
         }
     }
 }
+#endif
