+++ MADE.UI.Controls/Extensions/Extensions.Windows.Controls.cs
@@ -1,9 +1,10 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+﻿#if WINDOWS_UWP
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="Extensions.Controls.cs" company="MADE Apps">
 //   Copyright (c) MADE Apps.
 // </copyright>
 // <summary>
-//   Defines a collection of extensions for Android controls.
+//   Defines a collection of extensions for UWP controls.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
@@ -11,10 +12,11 @@ namespace MADE.UI.Controls
 {
     using System;
 
-    using Android.Views;
+    using Windows.UI.Xaml;
+    using Windows.UI.Xaml.Controls;
 
     /// <summary>
-    /// Defines a collection of extensions for Android controls.
+    /// Defines a collection of extensions for UWP controls.
     /// </summary>
     public static partial class Extensions
     {
@@ -27,7 +29,7 @@ namespace MADE.UI.Controls
         /// <param name="isVisible">
         /// A value indicating whether the control is visible.
         /// </param>
-        public static void SetVisible(this View view, bool isVisible)
+        public static void SetVisible(this UIElement view, bool isVisible)
         {
             SetVisible(view, isVisible, false);
         }
@@ -45,25 +47,24 @@ namespace MADE.UI.Controls
         /// <param name="updateChildViews">
         /// A value indicating whether to update the control's child visibility states.
         /// </param>
-        public static void SetVisible(this View view, bool isVisible, bool updateChildViews)
+        public static void SetVisible(this UIElement view, bool isVisible, bool updateChildViews)
         {
             if (view == null)
             {
                 return;
             }
 
-            view.Visibility = isVisible ? ViewStates.Visible : ViewStates.Gone;
+            view.Visibility = isVisible ? Visibility.Visible : Visibility.Collapsed;
 
-            if (!updateChildViews || !(view is ViewGroup viewGroup))
+            if (!updateChildViews || !(view is Panel panel))
             {
                 return;
             }
 
-            for (int i = 0; i < viewGroup.ChildCount; i++)
+            foreach (UIElement child in panel.Children)
             {
                 try
                 {
-                    View child = viewGroup.GetChildAt(i);
                     child?.SetVisible(isVisible, true);
                 }
                 catch (Exception)
@@ -74,3 +75,4 @@ namespace MADE.UI.Controls
         }
     }
 }
+#endif
