+++ MADE.UI.Controls/Extensions/Extensions.iOS.UIStackView.cs
@@ -1,4 +1,5 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+﻿#if __IOS__
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="Extensions.UIStackView.cs" company="MADE Apps">
 //   Copyright (c) MADE Apps.
 // </copyright>
@@ -38,3 +39,4 @@ namespace MADE.UI.Controls
         }
     }
 }
+#endif
