+++ MADE.UI.Controls/MADE.UI.Controls.csproj
@@ -1,7 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFrameworks>netstandard2.0;xamarinios10;monoandroid60;uap10.0.16299</TargetFrameworks>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <Authors>MADE Apps</Authors>
+    <Company>MADE Apps</Company>
+    <Product>MADE UI Controls Library</Product>
+    <Description>This package contains assemblies for helping to build custom controls in cross-platform applications built with the the .NET framework with support for UWP, Android, iOS and Xamarin.Forms applications.</Description>
+    <Copyright>Copyright (c) MADE Apps.</Copyright>
+    <PackageLicenseUrl>https://github.com/MADE-Apps/MADE-App-Components/blob/master/LICENSE</PackageLicenseUrl>
+    <PackageProjectUrl>https://github.com/MADE-Apps/MADE-App-Components</PackageProjectUrl>
+    <PackageIconUrl>https://pbs.twimg.com/profile_images/927154020422160385/6HSRU36P_400x400.jpg</PackageIconUrl>
+    <PackageTags>MADE Apps UI Controls Common Extensions Helpers Components Collections Runtime NET UWP Android iOS Web</PackageTags>
+    <Version>1.0.0</Version>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -19,7 +31,15 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\MADE.Common\MADE.Common.csproj" />
     <ProjectReference Include="..\MADE.UI\MADE.UI.csproj" />
   </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Condition=" '$(TargetFramework)' == 'uap10.0.16299' " Include="Microsoft.NETCore.UniversalWindowsPlatform " Version="6.0.8" />
+    <PackageReference Include="MSBuild.Sdk.Extras" Version="1.2.2" />
+  </ItemGroup>
+
+  <Import Project="$(MSBuildSDKExtrasTargets)" Condition="Exists('$(MSBuildSDKExtrasTargets)')" />
 
 </Project>
