+++ MADE.UI/Resources/Values/Colors.xml
